+++ src/modules/tools/filter/interfaces/filter.api.js
@@ -10,6 +10,7 @@ import InterfaceGeojsonIntern from "./interface.geojson.intern.js";
 import InterfaceGeojsonExtern from "./interface.geojson.extern.js";
 import InterfaceStaIntern from "./interface.sta.intern.js";
 import InterfaceStaExtern from "./interface.sta.extern.js";
+import InterfaceVectorTilesIntern from "./interface.vectortiles.intern";
 
 /**
  * FilterApi is the api to use in vue environment. It encapsulates the filter interfaces.
@@ -36,7 +37,8 @@ export default class FilterApi {
                 geojsonIntern: new InterfaceGeojsonIntern(FilterApi.intervalRegister, {getFeaturesByLayerId: openlayerFunctions.getFeaturesByLayerId, isFeatureInMapExtent: openlayerFunctions.isFeatureInMapExtent, isFeatureInGeometry: openlayerFunctions.isFeatureInGeometry}),
                 geojsonExtern: new InterfaceGeojsonExtern(),
                 staIntern: new InterfaceStaIntern(FilterApi.intervalRegister, {getFeaturesByLayerId: openlayerFunctions.getFeaturesByLayerId, isFeatureInMapExtent: openlayerFunctions.isFeatureInMapExtent, isFeatureInGeometry: openlayerFunctions.isFeatureInGeometry}),
-                staExtern: new InterfaceStaExtern()
+                staExtern: new InterfaceStaExtern(),
+                vectortilesIntern: new InterfaceVectorTilesIntern(FilterApi.intervalRegister, {getFeaturesByLayerId: openlayerFunctions.getVectorTileFeaturesByLayerId, isFeatureInMapExtent: openlayerFunctions.isFeatureInMapExtent, isFeatureInGeometry: openlayerFunctions.isFeatureInGeometry})
             };
         }
     }
@@ -55,16 +57,18 @@ export default class FilterApi {
      * @param {String} layerId the layer id
      * @param {ol/Layer} layerModel the layer model
      * @param {Boolean} extern if true, the type given by layerModel is used, otherwise "tree" is used
+     * @param {String} collection the collection - needed for vectortiles
      * @param {Function} onerror a function(Error)
      * @returns {void}
      */
-    setServiceByLayerModel (layerId, layerModel, extern, onerror) {
+    setServiceByLayerModel (layerId, layerModel, extern, collection, onerror) {
         if (!isObject(layerModel)) {
             return;
         }
         const type = layerModel.get("typ").toLowerCase(),
             featureNS = layerModel.get("featureNS"),
             url = layerModel.get("url"),
+            vectorTilesBaseUrl = layerModel.get("baseOAFUrl"),
             featureType = layerModel.get("featureType");
 
         if (type === "wfs") {
@@ -120,6 +124,26 @@ export default class FilterApi {
             }
             else {
                 onerror(new Error("FilterApi.setServiceByLayerModel: Filtering sta extern is not supported."));
+            }
+        }
+        else if (type === "vectortile") {
+            if (!extern) {
+                if (!vectorTilesBaseUrl) {
+                    onerror(new Error("FilterApi.setServiceByLayerModel: VectorTiles layer must have set the 'baseOAFUrl' param."));
+                    return;
+                }
+                this.service = {
+                    type,
+                    extern,
+                    layerId,
+                    url: vectorTilesBaseUrl,
+                    collection,
+                    namespace: featureNS,
+                    limit: typeof layerModel.get("limit") === "undefined" ? 400 : layerModel.get("limit")
+                };
+            }
+            else {
+                onerror(new Error("FilterApi.setServiceByLayerModel: Filtering vectortiles extern is not supported."));
             }
         }
         else if (typeof onerror === "function") {
@@ -357,6 +381,12 @@ export default class FilterApi {
         }
         else if (type === "sensorthings" && service.extern) {
             return FilterApi.interfaces.staExtern;
+        }
+        else if (type === "vectortile" && !service.extern) {
+            return FilterApi.interfaces.vectortilesIntern;
+        }
+        else if (type === "vectortile" && service.extern) {
+            return FilterApi.interfaces.oafExtern;
         }
         else if (typeof onerror === "function") {
             onerror(new Error("FilterApi.getInterfaceByService: Unknown service type " + type));