+++ src/modules/tools/filter/utils/openlayerFunctions.js
@@ -53,6 +53,32 @@ function getFeaturesByLayerId (layerId) {
 
     return layer.get("layer").getSource().getFeatures();
 }
+/**
+ * Gets the features of a vector tile layer by layer id and the collection.
+ * @param {Number} layerId The layerId to get the model by.
+ * @param {String} collection The collection which the features are wanted for.
+ * @param {Function} onsuccess The callback function to call on success - returns {ol/render/Feature[]}.
+ * @returns {void}
+ */
+function getVectorTileFeaturesByLayerId (layerId, collection, onsuccess) {
+    if (!collection || typeof layerId === "undefined") {
+        return;
+    }
+    const layerModel = getLayerByLayerId(layerId);
+
+    layerModel.layer.getSource().once("featuresloadend", event => {
+        const renderedFeatures = event.features;
+
+        onsuccess(renderedFeatures.filter(feature => {
+            if (typeof feature?.getProperties !== "function" || feature.getProperties()?.layer !== collection) {
+                return false;
+            }
+            return true;
+        }));
+    });
+    layerModel.layer.setOpacity(0);
+    layerModel.showFeaturesByIds([], true);
+}
 
 /**
  * Returns the current browser extent.
@@ -212,6 +238,7 @@ export default {
     getMapProjection,
     createLayerIfNotExists,
     getFeaturesByLayerId,
+    getVectorTileFeaturesByLayerId,
     getLayerByLayerId,
     getCurrentExtent,
     isFeatureInMapExtent,